requirements.txt
google-cloud-aiplatform>=1.60.0
google-cloud-bigquery>=3.20.0
